+++ packages/connect-wallet-modal/src/constants/conflictChecks.ts
@@ -7,7 +7,6 @@ export type ConflictChecks = {
   Exodus: ConflictCheck;
   Coin98: ConflictCheck;
   MathWallet: ConflictCheck;
-  Gamestop: ConflictCheck;
   Coinbase: ConflictCheck;
   Xdefi: ConflictCheck;
   Trust: ConflictCheck;
@@ -18,7 +17,6 @@ export const CONFLICTS: ConflictChecks = {
   Exodus: [helpers.isExodusProvider, 'Exodus'],
   Coin98: [helpers.isCoin98Provider, 'Coin98'],
   MathWallet: [helpers.isMathWalletProvider, 'MathWallet'],
-  Gamestop: [helpers.isGamestopProvider, 'GameStop'],
   Coinbase: [helpers.isCoinbaseProvider, 'Coinbase'],
   Xdefi: [helpers.isXdefiProvider, 'XDEFI'],
   Trust: [helpers.isTrustProvider, 'Trust'],