+++ packages/wallets-icons/src/react/index.ts
@@ -1,6 +1,4 @@
-export { default as Brave } from '../svg/brave.svg';
 export { default as Coinbase } from '../svg/coinbase.svg';
-export { default as Gamestop } from '../svg/gamestop.svg';
 export { default as ImToken } from '../svg/imtoken.svg';
 export { default as Ledger } from '../svg/ledger.svg';
 export { default as LedgerInversion } from '../svg/ledger-inversion.svg';
@@ -13,6 +11,5 @@ export { default as MathWalletInversion } from '../svg/mathwallet-inversion.svg'
 export { default as MetaMask } from '../svg/metamask.svg';
 export { default as MetaMaskCircle } from '../svg/metamask-circle.svg';
 export { default as MetaMaskCircleInversion } from '../svg/metamask-circle-inversion.svg';
-export { default as OperaWallet } from '../svg/opera-wallet.svg';
 export { default as Trust } from '../svg/trust-circle.svg';
 export { default as XDEFI } from '../svg/xdefi-wallet.svg';