+++ packages/web3-react/src/constants/connectors.ts
@@ -8,11 +8,7 @@ export enum PROVIDER_NAMES {
   INJECTED = 'Injected',
   LEDGER_HQ_LIVE = 'Ledger Live',
   LEDGER = 'Ledger',
-  COIN98 = 'Coin98',
   MATH_WALLET = 'MathWallet',
-  BRAVE = 'Brave Wallet',
-  OPERA = 'Opera Crypto Wallet',
-  GAMESTOP = 'GameStop Wallet',
   XDEFI = 'XDEFI',
   ZERION = 'Zerion',
 }